Refactoring: Extract almost all structs from Application.cpp
